requirements.txt
fastapi
uvicorn
pydantic
motor
python-dotenv
pydantic[email]
jose
passlib
python-jose[cryptography]
